Fix extended macros to also work with ActionButtonUseKeyDown=1
Turns out `SECURE_ACTIONS.click` does not pass the down parameter, so
the actual button gets always clicked with an up click and fails to
execute its action with `ActionButtonUseKeyDown=1`. A secure down click
can only be made with slash command, so the trigger macro must click the
actual button directly.

To avoid having to repeat the `/click [btn:1]...` sequence on both the
original button for the `ActionButtonUseKeyDown=1` and on the proxy
button for `ActionButtonUseKeyDown=0`, we can also wrap the original
button's click handler to capture the button there and set the attribute
directly on the Restricted Environment owner.

Unwrap is not needed here, I only had it in my example to allow edit and
re-execute without reloading the whole UI.

Changed indentation to tabs only for consistency.